% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_params_phenotypes.R
\name{calibrate_params_phenotypes}
\alias{calibrate_params_phenotypes}
\title{Calibration coefficients to compute phenotypes}
\usage{
calibrate_params_phenotypes(
  X0,
  B0,
  h2,
  b2,
  otu_list,
  Nqtl_y,
  Notu_y = length(otu_list)
)
}
\arguments{
\item{X0}{Matrix of the primary generation genotypes, given from \code{\link[MoPBS:get.geno]{MoPBS::get.geno()}}. SNPs are in rows and individuals in columns.}

\item{B0}{Matrix of the primary generation microbiomes, as CLR abundances. OTUs are in rows and individuals in columns.}

\item{h2}{direct heritability value, between 0 and 1}

\item{b2}{microbiability value, between 0 and 1}

\item{otu_list}{List of causal OTUs for the phenotypes.}

\item{Nqtl_y}{Integer; number of causal SNPs for the phenotypes.}

\item{Notu_y}{Integer; number of causal OTUs for the phenotypes.}
}
\value{
A \code{list} of phenotype parameters such as alpha, omega, list of causal SNPs for the phenotypes, standard deviation (=1) of phenotypes
}
\description{
This function is part of the first step of \code{\link[=holo_simu]{holo_simu()}}, focused on calibrating phynotypes parameters before looping on the generations. All coefficents are rescale during the process to ensure that standard deviation = 1 and the variance of microbiote effect and genetic effect satisfy the target microbiability and direct heritability.
}
\examples{
library(magrittr)
n_ind <- 10
n_snp <- 50
n_otu <- 50

# Simulate a small genotype matrix 
set.seed(123)
X0 <- matrix(sample(0:2, n_snp * n_ind, replace = TRUE), nrow = n_snp, ncol = n_ind)

# Simulate a small microbiome counts matrix
B0_counts_table <- matrix(abs(rnorm(n_snp * n_ind, mean = 10, sd = 3)), nrow = n_otu, ncol = n_ind)

# Transform to relative abundances per individuals and then apply CLR transformation
B0_abund <- apply(B0_counts_table, 2, function(x) x/sum(x))
B0 <- compositions::clr(t(B0_abund)) \%>\% t()

# Randomly select causal OTUs
rownames(B0) <- paste0("OTU", 1:n_otu)
otu_list <- c(y = sample(rownames(B0), 10))


params <- suppressWarnings(RITHMS:::calibrate_params_phenotypes(X0 = X0,
                                      B0 = B0,
                                      h2 = 0.25,
                                      b2 = 0.25,
                                      otu_list = otu_list,
                                      Nqtl_y = 10,
                                      Notu_y = length(otu_list)))
str(params)
 
}
\seealso{
\code{\link[=compute_phenotypes]{compute_phenotypes()}}
}
