% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_current_microbiome.R
\name{compute_current_microbiome}
\alias{compute_current_microbiome}
\title{Compute microbiomes for all individuals of current generation gathering all relevant object already computed.}
\usage{
compute_current_microbiome(
  beta,
  current_genotypes,
  mother_microbiomes,
  mean_microbiome,
  noise = 0.1,
  taxa_scale,
  lambda = 0.5,
  dir = F,
  thetaX
)
}
\arguments{
\item{beta}{A \code{matrix} of taxa (rows) by SNPs (columns) representing the multiplicative effect of SNPs on taxa abundances. Typically the output of \code{\link[=compute_beta_matrix_cluster]{compute_beta_matrix_cluster()}}}

\item{current_genotypes}{A \code{matrix} of genotypes for the individuals of the current generation.}

\item{mother_microbiomes}{A \code{matrix} containing the CLR-transformed abundances from the microbiome of the current mothers.}

\item{mean_microbiome}{A \code{numeric} vector representing the average abundances of the microbiome within the population.}

\item{noise}{A \code{numeric} scalar indicating the standard deviation of the Gaussian noise to add during beta g construction \code{\link[=compute_beta_g]{compute_beta_g()}}(multiplicative genetic effect on taxa abundances).}

\item{taxa_scale}{A \code{numeric} scalar to scale the noise added to each OTUs abundance, see \code{\link[=compute_beta_g]{compute_beta_g()}}.}

\item{lambda}{A \code{numeric} between 0 and 1 controlling the relative contribution of the mother microbiome compared to the average microbiome.}

\item{dir}{A \code{logical};}

\item{thetaX}{A \code{matrix} representing the environmental effect (optional).}
}
\value{
A \code{matrix} of CLR-transformed abundances of the microbiome for each individual of the current generation.
}
\description{
Compute microbiomes for all individuals of current generation gathering all relevant object already computed.
}
\seealso{
\code{\link[=compute_mean_microbiome]{compute_mean_microbiome()}}
}
