% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_params_phenotypes.R
\name{compute_mean_microbiome}
\alias{compute_mean_microbiome}
\title{apply mean on rows for each taxa across population}
\usage{
compute_mean_microbiome(microbiome, dir = F, n_ind = NULL, ao, mix.params)
}
\arguments{
\item{dir}{bool}

\item{n_ind}{number of individual per generation}

\item{ao}{for dirichlet sampling parameter}

\item{mix.params}{Vector of two values giving the weights for the regularisation of the base population microbiome. mix.params\link{1} = weight for raw microbiome and mix.params\link{2} = weight for mean microbiome. DEFAULT = c(0.75,0.25)}
}
\description{
apply mean on rows for each taxa across population
}
