% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_params_phenotypes.R
\name{compute_phenotypes}
\alias{compute_phenotypes}
\title{Compute phenotype values based on generated object of current generation}
\usage{
compute_phenotypes(
  X,
  B,
  otu_list,
  qtl_list,
  beta_qtl,
  beta_otu,
  Nqtl_y,
  Notu_y,
  se
)
}
\description{
Compute phenotype values based on generated object of current generation
}
