% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionnal-functions.R
\docType{data}
\name{Deru}
\alias{Deru}
\title{Deru Dataset}
\format{
A matrix with 780 rows (individuals) and 1845 taxa, where each element represents the abundance of a specific taxon in a specific individual. Additionnally, there is an attribute "population" wich provides the genotypes of each individual, encoded as 0, 1 ,2.
}
\source{
\href{https://pmc.ncbi.nlm.nih.gov/articles/PMC9194801/}{Déru et al.} for microbiota data and \href{https://pmc.ncbi.nlm.nih.gov/articles/PMC7538339/}{Déru et al. 2020} for genotypes data.
}
\usage{
Deru
}
\description{
A subset of data from \href{https://pmc.ncbi.nlm.nih.gov/articles/PMC7538339/}{Déru et al. 2020} containing hologenomic data from 750 pigs fed a conventional diet. Genotypes are accessible thanks to the "population" attribute.
}
\examples{
# Access the dataset
data("Deru")

# View the first few rows of the taxa matrix
head(Deru)

# Access the population attribute to view genotypes
attr(Deru, "population")
}
\keyword{datasets}
