% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_effect_calibration.R
\name{gen_effect_calibration}
\alias{gen_effect_calibration}
\title{Calibration of genetic effect from founder population data}
\usage{
gen_effect_calibration(
  founder_object,
  taxa_assign_g,
  correlation = 0.5,
  effect.size = c(seq(0.1, 1, 0.1)),
  plot = T
)
}
\arguments{
\item{founder_object}{Output of \code{\link[=generate_founder]{generate_founder()}} or \code{\link[=read_input_data]{read_input_data()}} function}

\item{taxa_assign_g}{Factor vector giving cluster assignment for all taxa, typical output of assign_taxa()}

\item{correlation}{Correlation between taxa within the same cluster, value between 0 and 1, DEFAULT = 0.5}

\item{effect.size}{Vector giving the size of genetic effect to try}

\item{plot}{boolean, if plot generation is required, DEFAULT = TRUE}
}
\value{
A data.frame with three columns, giving the Taxa ID, the effect.size and the corresponding heritability
}
\description{
Calibration of genetic effect from founder population data
}
\examples{
data("Deru")
ToyData <- Deru
taxa_assign_g <- assign_taxa(founder_object = ToyData)
effect_size_vector <- c(seq(0.1,1, by = 0.2))
out_data <- gen_effect_calibration(founder_object = ToyData,
                                   taxa_assign_g = taxa_assign_g,
                                   correlation = 0.5,
                                   effect.size = effect_size_vector,
                                   plot = TRUE)
}
