% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_data.R
\name{generate_founder}
\alias{generate_founder}
\title{Formatting of ped and map files into haplotypes table and filter individuals}
\usage{
generate_founder(path = NULL, microbiote_matrix)
}
\description{
Formatting of ped and map files into haplotypes table and filter individuals
}
