% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microbiomes.R
\name{get_mean_phenotypes}
\alias{get_mean_phenotypes}
\title{Useful function to extract mean phenotype values from RITHMS output}
\usage{
get_mean_phenotypes(data)
}
\description{
Useful function to extract mean phenotype values from RITHMS output
}
