% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microbiomes.R
\name{get_microbiomes}
\alias{get_microbiomes}
\title{Useful functions to extract elements of interest}
\usage{
get_microbiomes(data, transpose = F, CLR = F)
}
\description{
Useful functions to extract elements of interest
}
