% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microbiomes.R
\name{get_om_beta_g}
\alias{get_om_beta_g}
\title{Useful function to extract omega beta G values from RITHMS output}
\usage{
get_om_beta_g(data)
}
\description{
Useful function to extract omega beta G values from RITHMS output
}
