% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microbiomes.R
\name{get_phenotypes}
\alias{get_phenotypes}
\title{Useful function to extract all phenotype values as data frame from RITHMS output}
\usage{
get_phenotypes(data)
}
\description{
Useful function to extract all phenotype values as data frame from RITHMS output
}
