% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phenotypes_value.R
\name{get_phenotypes_value}
\alias{get_phenotypes_value}
\title{Useful function to extract phenotype values from RITHMS output ("gq" by default)}
\usage{
get_phenotypes_value(data, value = "gq")
}
\arguments{
\item{data}{List corresponding to one generation, as returned by \code{\link[=holo_simu]{holo_simu()}}. Containing simulation output.}
}
\value{
A \code{data.frame} of phenotype values for each individuals of a given generation.
}
\description{
The gets functions use the output of \code{\link[=holo_simu]{holo_simu()}} to extract the information of interest from a given generation.
\code{get_phenotypes_values} extract the set of phenotype values from a generation object.
}
\examples{
library(magrittr)
library(purrr)
datafile <- system.file("DeruPop.rds", package = "RITHMS")
ToyData <- readRDS(datafile)
taxa_assign_g <- assign_taxa(founder_object = ToyData)
generations_simu <- holo_simu(h2 = 0.25, b2 = 0.25, founder_object = ToyData,
                              n_clust = taxa_assign_g, n_ind = 500,
                              verbose = FALSE, seed = 1234)
                              
 #Extract phenotype values for each individuals of G1 generation
 G1_phenotypes_values <- get_phenotypes_value(generations_simu$G1)
 
 #Extract phenotype values for each individuals of all generations
 ## Don't forget to substract the metadata
 phenotypes_values <- generations_simu[-1] \%>\% map(get_phenotypes_value)
 
}
\seealso{
\code{\link[=get_mean_phenotypes]{get_mean_phenotypes()}}, \code{\link[=get_microbiomes]{get_microbiomes()}}, \code{\link[=get_om_beta_g]{get_om_beta_g()}}, \code{\link[=get_selected_ind]{get_selected_ind()}}, \code{\link[=get_phenotypes]{get_phenotypes()}}
}
