% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selected_ind.R
\name{get_selected_ind}
\alias{get_selected_ind}
\title{Useful function to extract ID of selected individuals from RITHMS output}
\usage{
get_selected_ind(data)
}
\arguments{
\item{data}{List corresponding to one generation, as returned by \code{\link[=holo_simu]{holo_simu()}}. Containing simulation output.}
}
\value{
A\code{list} of the selected individuals of a given generation.
}
\description{
The gets functions use the output of \code{\link[=holo_simu]{holo_simu()}} to extract the information of interest from a given generation.
\code{get_selected_ind}extract selectd individuals IDs from a generation object.
}
\examples{
\dontrun{
library(magrittr)
library(purrr)
datafile <- system.file("DeruPop.rds", package = "RITHMS")
ToyData <- readRDS(datafile)
taxa_assign_g <- assign_taxa(founder_object = ToyData)
generations_simu <- holo_simu(h2 = 0.25, b2 = 0.25, founder_object = ToyData,
                              n_clust = taxa_assign_g, n_ind = 500,
                              verbose = FALSE, seed = 1234)
                              
#Extract selected individuals IDs for the G1 generation
G1_selected_ind <- get_selected_ind(generations_simu$G1)
 
#Extract selected individuals IDs for all generations
## Don't forget to substract the metadata
selected_ind <- generations_simu[-1] \%>\% map(get_selected_ind)
}

}
\seealso{
\code{\link[=get_mean_phenotypes]{get_mean_phenotypes()}}, \code{\link[=get_phenotypes_value]{get_phenotypes_value()}}, \code{\link[=get_om_beta_g]{get_om_beta_g()}}, \code{\link[=get_microbiomes]{get_microbiomes()}}, \code{\link[=get_phenotypes]{get_phenotypes()}}
}
