% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holo_simu.R
\name{holo_simu}
\alias{holo_simu}
\title{Main function to generate transgenerational hologenomic data}
\usage{
holo_simu(
  h2,
  b2,
  founder_object,
  otu_g = 0.05,
  qtn_y = NULL,
  n_ind = NULL,
  n_gen = 5,
  n_clust = NULL,
  selection = F,
  size_selection_F = NULL,
  size_selection_M = NULL,
  selection_type = "GB",
  seed = 1234,
  verbose = T,
  lambda = 0.5,
  correlation = 0.5,
  mix.params = c(0.75, 0.25),
  mix.params.M = c(0.75, 0.25),
  effect.size = 0.1,
  noise.microbiome = 0.1,
  dir = F,
  ao = 25,
  thetaX = NULL,
  env_gen = NULL,
  w.param = c(0.5, 0.5)
)
}
\arguments{
\item{h2}{direct heritability value, between 0 and 1}

\item{b2}{microbiability value, between 0 and 1}

\item{founder_object}{output of generate_founder() function}

\item{otu_g}{percentage of taxa under genetic control, between 0 and 1, DEFAULT = 0.05}

\item{qtn_y}{number of causal SNPs for the phenotypes}

\item{n_ind}{number of individual per generation}

\item{n_gen}{number of generation, DEFAULT = 5}

\item{n_clust}{vector with taxa assignment, typical output of assign_taxa}

\item{selection}{bool, if selection process needed, DEFAULT = FALSE}

\item{size_selection_F}{percentage of female to select}

\item{size_selection_M}{percentage of male to select}

\item{selection_type}{mode of selection to be used, value in ("GB", "B", "G", "diversity", "div.GB"), DEFAULT = "GB"}

\item{seed}{seed value for samplings in the function}

\item{verbose}{bool, DEFAULT = T}

\item{lambda}{proportion of microbiome of offspring coming from vertical transmission, DEFAULT = 0.5}

\item{correlation}{Correlation between taxa within the same cluster, value between 0 and 1, DEFAULT = 0.5}

\item{mix.params}{Vector of two values giving the weights for the regularisation of the base population microbiome. mix.params\link{1} = weight for raw microbiome and mix.params\link{2} = weight for mean microbiome. DEFAULT = c(0.75,0.25)}

\item{noise.microbiome}{sd of microbiome noise, DEFAULT = 0.1}

\item{dir}{bool}

\item{ao}{for dirichlet sampling parameter}

\item{thetaX}{if environmental effect}

\item{env_gen}{vector of bool}

\item{w.param}{in case div.GB selection mode is chosen}
}
\value{
A big list object with metada info such as beta matrix details and each generation at level 1. For each generation, the genotypes, the microbiomes, the phenotypes, the pedigree and the individuals selected can be reachable.
}
\description{
Main function to generate transgenerational hologenomic data
}
\examples{
\dontrun{
  datafile <- system.file("DeruPop.rds", package = "RITHMS")
  ToyData <- readRDS(datafile)
  taxa_assign_g <- assign_taxa(founder_object = ToyData)
  generations_simu <- holo_simu(h2 = 0.25,
                                 b2 = 0.25,
                                 founder_object = ToyData,
                                 n_clust = taxa_assign_g,
                                 n_ind = 500,
                                 verbose = FALSE,
                                 noise.microbiome = 0.5,
                                 effect.size = 0.3,
                                 lambda = 0.5,
                                 dir = TRUE,
                                 selection = FALSE,
                                 seed = 1234)
}  
}
\seealso{
\code{\link[=compute_beta_matrix_cluster]{compute_beta_matrix_cluster()}}, \code{\link[=compute_mean_microbiome]{compute_mean_microbiome()}}, \code{\link[=compute_current_microbiome]{compute_current_microbiome()}}, \code{\link[=compute_phenotypes]{compute_phenotypes()}}
}
