% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_data.R
\name{rarefied_microbiome}
\alias{rarefied_microbiome}
\title{Rarefaction step within read_input_data}
\usage{
rarefied_microbiome(microbiome_matrix, threshold = 0.05)
}
\arguments{
\item{threshold}{Threshold for rarefaction, DEFAULT = 0.05}
}
\description{
Rarefaction step within read_input_data
}
