% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input_data.R
\name{read_input_data}
\alias{read_input_data}
\title{Formatting data from file paths to base population object}
\usage{
read_input_data(
  path_to_microbiome,
  path_to_genotype,
  file_type = "pedmap",
  biome_id_column = "ind_id",
  threshold = 0.05,
  ind_selected = NULL
)
}
\arguments{
\item{path_to_microbiome}{String giving the path to count table file. \code{"path/to/microbiome.{extension}"}}

\item{path_to_genotype}{String giving the path and prefix to the genotype file. This should be one of \code{"pedmap"} or \code{"vcf"}. (default: \code{"pedmap"}). For a \code{"example_pedmap.ped"} file enter only \code{"path/to/example_pedmap"}.}

\item{biome_id_column}{String specifying the name of the column containing the individual IDs for the microbiome matrix. Default is "ind_id".}

\item{threshold}{Threshold for rarefaction, DEFAULT = 0.05}

\item{ind_selected}{Vector of string values with individuals to keep, have to match rownames of count table file, DEFAULT = NULL}
}
\value{
A data.frame corresponding to (rarefied) microbiome with individuals in rows and taxa in columns. Genotypes data.frame is an attribute called "population" reachable using \code{attr(output_name,"population")}
}
\description{
Formatting data from file paths to base population object
}
\note{
The number of individuals in the microbiome an genotype data must be equivalent.
}
\examples{
\dontrun{
# Create founder object from PED/MAP set
founder_object <- read_input_data(path_to_microbiome = "/path/to/microbiome.txt",
                                  path_to_pedmap = "/path/to/pedmap/'prefix'",
                                  biome_id_column = "ind_id",
                                  threshold = 0.05,
                                  ind_selected = NULL,
                                  file_type = "pedmap")

# Create founder object from VCF
founder_object <- read_input_data(path_to_microbiome = "/path/to/microbiome.txt",
                                  path_to_pedmap = "/path/to/vcf/'prefix'",
                                  biome_id_column = "ind_id",
                                  threshold = 0.05,
                                  ind_selected = NULL,
                                  file_type = "vcf")
}
}
\seealso{
\code{\link[=generate_founder]{generate_founder()}}, \code{\link[=rarefied_microbiome]{rarefied_microbiome()}}
}
