% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_params_phenotypes.R
\name{replace_zero}
\alias{replace_zero}
\title{Replace zero in the matrix by 1e-15 to avoid infinite values when CLR transformation is computed}
\usage{
replace_zero(x)
}
\description{
Replace zero in the matrix by 1e-15 to avoid infinite values when CLR transformation is computed
}
