% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_params_phenotypes.R
\name{richness_from_abundances_gen}
\alias{richness_from_abundances_gen}
\title{If selection based on diversity, need to go back to counts using multinomial sampling from abundances.}
\usage{
richness_from_abundances_gen(microbiome_matrix, n_loop = 10, plot = T)
}
\description{
If selection based on diversity, need to go back to counts using multinomial sampling from abundances.
}
