% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{root}
\alias{root}
\title{Compute root value for beta matrix construction}
\usage{
root(rho)
}
\arguments{
\item{rho}{A numeric value between 0 and 1 representing the target correlation between taxa within the same cluster. This parameter is used to induce correlation in the beta matrix.}
}
\description{
This function computes the parameter used to build a beta matrix with a given correlation structure between taxa. It is used in the beta matrix construction.
}
