% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_params_phenotypes.R
\name{select_individual}
\alias{select_individual}
\title{Obtain id of the parents selected for the next generation based on the criteria chosen by the user}
\usage{
select_individual(
  phenotypes,
  microbiomes,
  genotypes,
  beta,
  beta_otu,
  selection,
  size_selection_F,
  size_selection_M,
  selection_type,
  w.param
)
}
\arguments{
\item{selection}{bool, if selection process needed, DEFAULT = FALSE}

\item{size_selection_F}{percentage of female to select}

\item{size_selection_M}{percentage of male to select}

\item{selection_type}{mode of selection to be used, value in ("GB", "B", "G", "diversity", "div.GB"), DEFAULT = "GB"}

\item{w.param}{in case div.GB selection mode is chosen}
}
\description{
Obtain id of the parents selected for the next generation based on the criteria chosen by the user
}
