% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\docType{data}
\name{taxonomy}
\alias{taxonomy}
\title{Taxonomy assignation}
\format{
A \code{data.frame} containing the taxonomy for each OTU sequenced matching the ID in the microbiota matrix till Kingdom level.
}
\source{
\href{https://pmc.ncbi.nlm.nih.gov/articles/PMC9194801/}{Déru et al.} for microbiota data and corresponding metadata
}
\usage{
taxonomy
}
\description{
Taxonomy assignation
}
\keyword{datasets}
