% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_geno_into_vcf.R
\name{transform_geno_into_vcf}
\alias{transform_geno_into_vcf}
\title{Convert a 0/1/2 genotype matrix into a VCF-like format}
\usage{
transform_geno_into_vcf(
  geno_matrix,
  output_type = c("file", "dataframe", "both"),
  output_path = NULL
)
}
\arguments{
\item{geno_matrix}{A matrix of genotypes with values 0, 1, 2. SNPs are in rows and individuals in columns.}

\item{output_type}{A character, that specifies the output type. Choose between \code{"file"} (write \code{.vcf}),
\code{"dataframe"} (return \code{data.frame}), or \code{"both"} (do both).}

\item{output_path}{A character string that specifies the output path. Requiered if \code{output_type} is "file" or "both".}
}
\value{
A \code{data.frame} if \verb{output_type = "dataframe" or "both"}, or just write the \code{.vcf} file if \verb{output_type = "file" or "both"}.
}
\description{
This function converts a genotype matrix encoded as 0,1,2 into a VCF-like format. It can either return the VCF content as a \code{data.frame}, write it to a .vcf file or do both.
}
\examples{
n_ind <- 50
n_snp <- 50

# Create a small genotype matrix
set.seed(123)
geno_matrix <- matrix(sample(0:2, size = n_ind*n_snp, replace = TRUE),
                      nrow = n_snp, ncol = n_ind)

# Transform genotype matrix into VCF
vcf <- transform_geno_into_vcf(geno_matrix, output_type = "dataframe")
head(vcf[1:3,1:7])
}
