% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_data.R
\name{compute_beta_g}
\alias{compute_beta_g}
\title{Compute product of matrixes based on few parameters}
\usage{
compute_beta_g(beta, genotypes, noise, taxa_scale)
}
\description{
Compute product of matrixes based on few parameters
}
