% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_beta_matrix_cluster.R
\name{compute_beta_matrix_cluster}
\alias{compute_beta_matrix_cluster}
\title{Generate beta matrix giving genetic effect per SNP on taxa abundances.}
\usage{
compute_beta_matrix_cluster(
  n_b,
  n_g,
  n_clust,
  n_qtl_o,
  n_otus,
  effect_size = 1,
  correlation = 1,
  beta_info = NULL
)
}
\arguments{
\item{n_b}{Number of taxa}

\item{n_g}{Number of SNPs}

\item{n_clust}{A vector with a length matching the total number of taxa with values from 0 to the number of clusters. Typically the \code{\link[=assign_taxa]{assign_taxa()}} output.}

\item{n_qtl_o}{Number of causative QTL on taxa abundances (per taxon)}

\item{n_otus}{Number of taxa under genetic control.}

\item{effect_size}{Vector giving the size of genetic effect to try}

\item{correlation}{A numeric value between 0 and 1 representing the target correlation between taxa within the same cluster.}

\item{beta_info}{Informations from Beta matrix of genetic effects}
}
\value{
A \code{list} of two objects linked to the beta matrix.
The beta matrix itself : a \code{matrix} of taxa (in rows) across SNPs (in columns) giving the multiplicative effect of genotype on taxa abundances.
A small \code{data.frame} containing the parameters used during the beta matrix simulation, including \code{cluster}, \code{id_otu} and \code{id_qtl_o}.
}
\description{
Generate beta matrix giving genetic effect per SNP on taxa abundances.
}
\seealso{
\code{\link[=compute_beta_g]{compute_beta_g()}}, \code{\link[=holo_simu]{holo_simu()}}
}
