% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_data.R
\name{compute_beta_matrix_cluster}
\alias{compute_beta_matrix_cluster}
\title{Generate beta matrix giving genetic effect per SNP on taxa abundances.}
\usage{
compute_beta_matrix_cluster(
  n_b,
  n_g,
  n_clust,
  n_qtl_o,
  n_otus,
  effect_size = 1,
  correlation = 1,
  beta_info = NULL
)
}
\arguments{
\item{correlation}{Correlation between taxa within the same cluster, value between 0 and 1, DEFAULT = 0.5}
}
\description{
Generate beta matrix giving genetic effect per SNP on taxa abundances.
}
