% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_params_phenotypes.R
\name{compute_current_microbiome}
\alias{compute_current_microbiome}
\title{Compute microbiomes for all individuals of current generation gathering all relevant object already computed.}
\usage{
compute_current_microbiome(
  beta,
  current_genotypes,
  mother_microbiomes,
  mean_microbiome,
  noise = 0.1,
  taxa_scale,
  lambda = 0.5,
  dir = F,
  thetaX
)
}
\arguments{
\item{lambda}{proportion of microbiome of offspring coming from vertical transmission, DEFAULT = 0.5}

\item{dir}{bool}

\item{thetaX}{if environmental effect}
}
\description{
Compute microbiomes for all individuals of current generation gathering all relevant object already computed.
}
