% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mean_microbiome.R
\name{compute_mean_microbiome}
\alias{compute_mean_microbiome}
\title{Compute the mean for each taxa across population}
\usage{
compute_mean_microbiome(microbiome, dir = F, n_ind = NULL, ao, mix.params)
}
\arguments{
\item{microbiome}{A \code{matrix} of microbiome abundances (taxa in rows, individuals in columns).}

\item{dir}{A \code{logical}; if \code{TRUE}, use a Dirichlet distribution to simulate inter-individual variability. Default is \code{FALSE}.}

\item{n_ind}{The number of individuals to simulate with the Dirichlet distribution (required if \code{dir = TRUE}).}

\item{ao}{A numeric scalar used as the concentration parameter for the Dirichlet distribution.}

\item{mix.params}{A numeric vector (\code{length = 2}) specifying weights between Dirichlet samples and the original mean. The weigths should sum to 1.}
}
\value{
A \code{numeric}vector of mean microbiome values if \code{dir = FALSE}.
A \code{matrix} of simulated microbiomes if \code{dir = TRUE}
}
\description{
This function estimate the mean microbiome by averaging abundances across individuals for each taxa. Optionally,
it use a Dirichlet distribution to simulate inter-individual variability centered arround the mean.
}
\seealso{
\code{\link[=compute_current_microbiome]{compute_current_microbiome()}}
}
