% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_phenotypes.R
\name{compute_phenotypes}
\alias{compute_phenotypes}
\title{Compute phenotype values based on generated objects of the current generation}
\usage{
compute_phenotypes(
  X,
  B,
  otu_list,
  qtl_list,
  beta_qtl,
  beta_otu,
  Nqtl_y,
  Notu_y,
  se
)
}
\arguments{
\item{X}{A \code{matrix} of the current generation genotypes, encoding with 0, 1 and 2.}

\item{B}{A \code{matrix} of the CLR-transformed abundances of the current generation microbiome.}

\item{otu_list}{List of causal OTUs for the phenotypes.}

\item{qtl_list}{List of causal SNPs for the phenotypes.}

\item{beta_qtl}{Alpha, the regression coefficients corresponding to the QTL effects on the phenotype. See \code{\link[=calibrate_params_phenotypes]{calibrate_params_phenotypes()}}.}

\item{beta_otu}{Omega, the regression coefficients corresponding to taxa effects on the phenotype. See \code{\link[=calibrate_params_phenotypes]{calibrate_params_phenotypes()}}.}

\item{Nqtl_y}{Integer, number of causal SNPs for the phenotypes.}

\item{Notu_y}{Integer, number of causal OTUs for the phenotypes.}

\item{se}{Phenotypes Standard deviation, see \code{\link[=calibrate_params_phenotypes]{calibrate_params_phenotypes()}}.}
}
\description{
Compute phenotype values based on generated objects of the current generation
}
\seealso{
\code{\link[=calibrate_params_phenotypes]{calibrate_params_phenotypes()}}
}
