% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_data.R
\name{assign_taxa}
\alias{assign_taxa}
\title{Assign all taxa to a cluster, eventually under genetic control using hclust}
\usage{
assign_taxa(founder_object, taxa_g = 0.05)
}
\arguments{
\item{founder_object}{Output of read_input_data function}

\item{taxa_g}{Percentage of taxa under genetic control, DEFAULT = 0.1}
}
\value{
A vector with a length matching the total number of taxa with values from 0 to the number of taxa, 0 corresponding to the non under genetic control cluster
}
\description{
Assign all taxa to a cluster, eventually under genetic control using hclust
}
\examples{
datafile <- system.file("DeruPop.rds", package = "RITHMS")
ToyData <- readRDS(datafile)
assign_taxa(founder_object = ToyData,
            taxa_g = 0.2)
}
