% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_founder.R
\name{generate_founder}
\alias{generate_founder}
\title{Formatting of ped and map files into haplotypes table and filter individuals}
\usage{
generate_founder(path = NULL, microbiome_matrix, file_type = "pedmap")
}
\arguments{
\item{path}{String giving the path and prefix to ped and map file}

\item{microbiome_matrix}{Filtered microbiome matrix, with OTUs in columns and individuals in rows}

\item{file_type}{String specifying the file type used to load genotype data. This should be one of \code{"pedmap"} or \code{"vcf"}. (default: \code{"pedmap"}).}
}
\value{
Matrix of taxa (in columns) accross individuals (in rows), genotypes of each generation are reachable thanks to \strong{"population"} attribute. Genotypes are encodes as 0,1,2 and generated using the \href{https://pubmed.ncbi.nlm.nih.gov/32229505/}{MoBPS} package.
}
\description{
Formatting of ped and map files into haplotypes table and filter individuals
}
\examples{
\dontrun{
  microbiome <- data.table::fread("/path/to/microbiome.txt")
  founder_object <- generate_founder(path = "/path/to/pedmap", microbiome) 
}

}
\seealso{
\code{\link[=read_input_data]{read_input_data()}}, \code{\link[=rarefied_microbiome]{rarefied_microbiome()}}
}
