% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microbiomes.R
\name{get_mean_diversity}
\alias{get_mean_diversity}
\title{Useful function to extract Shannon diversity from RITHMS output}
\usage{
get_mean_diversity(data)
}
\description{
Useful function to extract Shannon diversity from RITHMS output
}
