% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_phenotypes.R
\name{get_mean_phenotypes}
\alias{get_mean_phenotypes}
\title{Useful function to extract mean phenotype values from RITHMS output}
\usage{
get_mean_phenotypes(data)
}
\arguments{
\item{data}{List corresponding to one generation, as returned by \code{\link[=holo_simu]{holo_simu()}}. Containing simulation output.}
}
\value{
The mean phenotype value for a given generation.
}
\description{
The gets functions use the output of \code{\link[=holo_simu]{holo_simu()}} to extract the information of interest from a given generation.
\code{get_mean_phenotypes}extract average phenotype from a generation object.
}
\examples{
\dontrun{
library(purrr)
library(magrittr)
data("Deru")
ToyData <- Deru
taxa_assign_g <- assign_taxa(founder_object = ToyData)
generations_simu <- holo_simu(h2 = 0.25, b2 = 0.25, founder_object = ToyData,
                              n_clust = taxa_assign_g, n_ind = 500,
                              verbose = FALSE, seed = 1234)
                              
# Extract mean phenotype value for G1 generation
G1_mean_phenotype <- get_mean_phenotypes(generations_simu$G1)

# Extract mean phenotype values of all generations
## Don't forget to substract the metadata
mean_phenotypes <- generations_simu[-1] \%>\% map(get_mean_phenotypes)
}

}
\seealso{
\code{\link[=get_microbiomes]{get_microbiomes()}}, \code{\link[=get_phenotypes_value]{get_phenotypes_value()}}, \code{\link[=get_om_beta_g]{get_om_beta_g()}}, \code{\link[=get_selected_ind]{get_selected_ind()}}, \code{\link[=get_phenotypes]{get_phenotypes()}}
}
