% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microbiomes.R
\name{get_microbiomes}
\alias{get_microbiomes}
\title{Useful function to extract microbiomes from RITHMS output}
\usage{
get_microbiomes(data, transpose = F, CLR = F)
}
\arguments{
\item{data}{List corresponding to one generation, as returned by \code{holo_simu()}. Containing simulation output.}

\item{transpose}{Logical; if \code{TRUE}, transpose the microbiome matrix (OTUs in rows, individuals in columns).}

\item{CLR}{Logical; if \code{TRUE}, applies a CLR transformation to the abundance data. This transformation requires \code{transpose = TRUE}.}
}
\value{
A \code{data.frame}containing the microbiome abundances of individuals. Default, individuals are in rows and OTUs in columns. Change \code{transpose}parameter if needed.
}
\description{
The gets functions use the output of \code{\link[=holo_simu]{holo_simu()}} to extract the information of interest from a given generation.
\code{get_microbiomes} extract the microbiome abundance matrix from a generation object, with or without CLR transformation or transposition.
}
\examples{
\dontrun{
library(magrittr)
library(purrr)
data("Deru")
ToyData <- Deru
taxa_assign_g <- assign_taxa(founder_object = ToyData)
generations_simu <- holo_simu(h2 = 0.25, b2 = 0.25, founder_object = ToyData,
                              n_clust = taxa_assign_g, n_ind = 500,
                              verbose = FALSE, seed = 1234)

# Extract microbiome matrix for G1 generation
G1_microbiome <- get_microbiomes(generations_simu$G1)

# Extract with transposition
G1_t_microbiome <- get_microbiomes(generations_simu$G1, transpose = TRUE)

# Extract with CLR transformation
G1_CLR_microbiome <- get_microbiomes(generations_simu$G1, transpose = TRUE, CLR = TRUE)

# Extract all microbiome matrices of all generations
# substract metadata
microbiomes <- generations_simu[-1] \%>\% map(get_microbiomes)
}

}
\seealso{
\code{\link[=get_mean_phenotypes]{get_mean_phenotypes()}}, \code{\link[=get_phenotypes_value]{get_phenotypes_value()}}, \code{\link[=get_om_beta_g]{get_om_beta_g()}}, \code{\link[=get_selected_ind]{get_selected_ind()}}, \code{\link[=get_phenotypes]{get_phenotypes()}}
}
