% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microbiomes.R
\name{get_microbiomes}
\alias{get_microbiomes}
\title{Useful function to extract microbiomes from RITHMS output}
\usage{
get_microbiomes(data, transpose = F, CLR = F)
}
\description{
Useful function to extract microbiomes from RITHMS output
}
