% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microbiomes.R
\name{get_phenotypes_value}
\alias{get_phenotypes_value}
\title{Useful function to extract phenotype values from RITHMS output ("gq" by default)}
\usage{
get_phenotypes_value(data, value = "gq")
}
\description{
Useful function to extract phenotype values from RITHMS output ("gq" by default)
}
