% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microbiomes.R
\name{get_selected_ind}
\alias{get_selected_ind}
\title{Useful function to extract ID of selected individuals from RITHMS output}
\usage{
get_selected_ind(data)
}
\description{
Useful function to extract ID of selected individuals from RITHMS output
}
