% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holo_simu.R
\name{holo_simu}
\alias{holo_simu}
\title{Main function to generate transgenerational hologenomic data}
\usage{
holo_simu(
  h2,
  b2,
  founder_object,
  n_ind = NULL,
  n_clust = NULL,
  n_gen = 5,
  qtn_y = NULL,
  correlation = 0.5,
  otu_g = 0.05,
  lambda = 0.5,
  effect.size = 0.1,
  mix.params = c(0.75, 0.25),
  mix.params.M = c(0.75, 0.25),
  noise.microbiome = 0.1,
  dir = T,
  ao = 25,
  size_rmultinom = 10000,
  selection = F,
  size_selection_F = NULL,
  size_selection_M = NULL,
  selection_type = "GB",
  w.param = c(0.5, 0.5),
  thetaX = NULL,
  env_gen = NULL,
  seed = 1234,
  verbose = T
)
}
\arguments{
\item{h2}{direct heritability value, between 0 and 1.}

\item{b2}{microbiability value, between 0 and 1.}

\item{founder_object}{output of \code{generate_founder()} function.}

\item{n_ind}{number of individual per generation.}

\item{n_clust}{vector with taxa assignment, typical output of \code{assign_taxa()}}

\item{n_gen}{number of generation, DEFAULT = 5}

\item{qtn_y}{number of causal SNPs for the phenotypes.}

\item{correlation}{Correlation between taxa within the same cluster, value between 0 and 1, DEFAULT = 0.5}

\item{otu_g}{percentage of taxa under genetic control, between 0 and 1, DEFAULT = 0.05}

\item{lambda}{proportion of microbiome of offspring coming from vertical transmission, DEFAULT = 0.5}

\item{effect.size}{Vector giving the size of genetic effect to try.}

\item{mix.params}{Vector of two numeric values giving the weights for the regularisation of the base population microbiome. \code{mix.params[1]} = weight for raw microbiome and \code{mix.params[2]} = weight for mean microbiome. DEFAULT = c(0.75,0.25).}

\item{mix.params.M}{A vector of two numeric values specifying weights between Dirichlet samples and the original mean. DEFAULT = c(0.75,0.25). With \code{mix.params.M[1]} the dirichlet microbiome coefficient and \code{mix.params.M[2]} the mean microbiome coefficient.}

\item{noise.microbiome}{sd of microbiome noise, DEFAULT = 0.1}

\item{dir}{Logical; Mentions if the ambient microbiome is generated via a Dirichlet law or a only the \code{mean_microbiome}. DEFAULT = T}

\item{ao}{A numeric scalar used as the concentration parameter for the Dirichlet distribution.}

\item{size_rmultinom}{Integer; specifying the total number of object for the multinomial sampling(default: 10000, according to DeruPop.rds dataset).}

\item{selection}{bool, if selection process needed, DEFAULT = FALSE}

\item{size_selection_F}{percentage of female to select.}

\item{size_selection_M}{percentage of male to select.}

\item{selection_type}{mode of selection to be used, value in ("GB", "B", "G", "diversity", "div.GB"), DEFAULT = "GB"}

\item{w.param}{in case div.GB selection mode is chosen.}

\item{thetaX}{Optionnal matrix specifying environmental effects applied to the microbiome, such as antibiotic treatment. This matrix should be of dimension \verb{n_taxa x n_individuals}, and typically constructed as the product of:
\itemize{
\item a vector of taxa-specific effects \code{theta} (negative or positive values for all taxa).
\item a binary vector \code{X} encoding individual exposure (1 for treated, 0 ofr untreated).
If \code{NULL} (default value), no environmental effect is applied.
Example use case: refer to the vignette on \href{https://solenepety.github.io/RITHMS/articles/generate-figures.html#introduction-of-transient-perturbations-of-the-microbiota}{Generate figures}
}}

\item{env_gen}{vector of booleans.}

\item{seed}{seed value for samplings in the function.}

\item{verbose}{bool, DEFAULT = T}
}
\value{
A big list object with metada info such as beta matrix details and each generation at level 1. For each generation, the genotypes, the microbiomes, the phenotypes, the pedigree and the individuals selected can be reachable.
}
\description{
Simulates hologenomic data over multiple generations with genetic, microbial, and environmental effects.
}
\section{Required parameters}{

\itemize{
\item \code{h2}
\item \code{b2}
\item \code{founder_object}
\item \code{n_ind}
\item \code{n_clust}
}
}

\section{Generation parameters}{

\itemize{
\item \code{n_gen}
}
}

\section{Genotype-related parameters}{

\itemize{
\item \code{qtn_y}
}
}

\section{Microbiome-related parameters}{

\itemize{
\item \code{correlation}
\item \code{otu_g}
\item \code{lambda}
\item \code{effect.size}
\item \code{mix.params}
\item \code{mix.params.M}
\item \code{noise.microbiome}
\item \code{dir}
\item \code{ao}
}
}

\section{Selection-related parameters}{

\itemize{
\item \code{selection}
\item \code{size_selection_F}, \code{size_selection_M}
\item \code{selection_type}
}
}

\section{Environmental effects}{

\itemize{
\item \code{thetaX}
\item \code{env_gen}
}
}

\section{Misc}{

\itemize{
\item \code{seed}
\item \code{verbose}
}
}

\examples{
\dontrun{
  data("Deru")
  ToyData <- Deru
  taxa_assign_g <- assign_taxa(founder_object = ToyData)
  generations_simu <- holo_simu(h2 = 0.25,
                                 b2 = 0.25,
                                 founder_object = ToyData,
                                 n_clust = taxa_assign_g,
                                 n_ind = 500,
                                 verbose = FALSE,
                                 noise.microbiome = 0.5,
                                 effect.size = 0.3,
                                 lambda = 0.5,
                                 dir = TRUE,
                                 selection = FALSE,
                                 seed = 1234)
}  
}
\seealso{
\code{\link[=compute_beta_matrix_cluster]{compute_beta_matrix_cluster()}}, \code{\link[=compute_mean_microbiome]{compute_mean_microbiome()}}, \code{\link[=compute_current_microbiome]{compute_current_microbiome()}}, \code{\link[=compute_phenotypes]{compute_phenotypes()}}
}
