% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefied_microbiome.R
\name{rarefied_microbiome}
\alias{rarefied_microbiome}
\title{Rarefaction step within read_input_data}
\usage{
rarefied_microbiome(microbiome_matrix, id_column = "ind_id", threshold = 0.05)
}
\arguments{
\item{microbiome_matrix}{Microbiome counting table, OTUs in columns and a column specifying the IDs}

\item{id_column}{String specifying the name of the column containing the individual IDs. Default is "ind_id"}

\item{threshold}{Minimum prevalence threshold}
}
\value{
The filtered microbiome matrix, without OTUs whose prevalence is lower than the set threshold
}
\description{
Rarefaction step within read_input_data
}
\examples{
\dontrun{
  microbiome <- data.table::fread("/path/to/microbiome/'prefix'")
  microbiome_filtered <- rarefied_microbiome(microbiome, id_columns = "sample_id", threshold = 0.05)
}
}
\seealso{
\code{\link[=read_input_data]{read_input_data()}}, \code{\link[=generate_founder]{generate_founder()}}
}
