% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input_data.R
\name{read_input_data}
\alias{read_input_data}
\title{Formatting data from file paths to base population object}
\usage{
read_input_data(
  path_to_microbiome,
  path_to_pedmap,
  biome_id_columns = "ind_id",
  threshold = 0.05,
  ind_selected = NULL
)
}
\arguments{
\item{path_to_microbiome}{String giving the path to count table file}

\item{path_to_pedmap}{String giving the path and prefix to ped and map file}

\item{threshold}{Threshold for rarefaction, DEFAULT = 0.05}

\item{ind_selected}{Vector of string values with individuals to keep, have to match rownames of count table file, DEFAULT = NULL}

\item{biome_id_column}{String specifying the name of the column containing the individual IDs for the microbiome matrix. Default is "ind_id".}
}
\value{
A data.frame corresponding to (rarefied) microbiome with individuals in rows and taxa in columns. Genotypes data.frame is an attribute called "population" reachable using \code{attr(output_name,"population")}
}
\description{
Formatting data from file paths to base population object
}
\examples{
# founder_object <- read_input_data(path_to_microbiome = "/path/to/microbiome/'prefix'",
#                                   path_to_pedmap = "/path/to/pedmap/'prefix'",
#                                   biome_id_column = "ind_id",
#                                   threshold = 0.05,
#                                   ind_selected = NULL)

}
