% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_current_microbiome.R
\name{replace_zero}
\alias{replace_zero}
\title{Replace zero in the matrix by 1e-15 to avoid infinite values when CLR transformation is computed}
\usage{
replace_zero(x)
}
\arguments{
\item{x}{A matrix}
}
\description{
Replace zero in the matrix by 1e-15 to avoid infinite values when CLR transformation is computed
}
