% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_individual.R
\name{select_individual}
\alias{select_individual}
\title{Obtain id of the parents selected for the next generation based on the criteria chosen by the user}
\usage{
select_individual(
  phenotypes,
  microbiomes,
  genotypes,
  beta,
  beta_otu,
  selection,
  size_selection_F,
  size_selection_M,
  selection_type,
  w.param
)
}
\arguments{
\item{phenotypes}{Phenotype values of the current generation given as the result of the combined effects of the microbiota and direct genetic effects. Typically \code{\link[=get_phenotypes]{get_phenotypes()}} output or \code{\link[=compute_phenotypes]{compute_phenotypes()}}.}

\item{microbiomes}{Abundances for each individual of the current generation. Note: It is necessary to transform the abundances if they are CLR-transformed with the following command. \code{microbiomes_clr |>  t() |> clrInv() |> t()}}

\item{genotypes}{Genotypes values of the current generation. Typically from the MoBPS \code{get.geno()} function.}

\item{beta}{Beta matrix from \code{\link[=compute_beta_matrix_cluster]{compute_beta_matrix_cluster()}} output.}

\item{beta_otu}{Omega parameter from \code{\link[=calibrate_params_phenotypes]{calibrate_params_phenotypes()}} output.}

\item{selection}{bool, if selection process needed, DEFAULT = FALSE}

\item{size_selection_F}{percentage of female to select}

\item{size_selection_M}{percentage of male to select}

\item{selection_type}{mode of selection to be used, value in ("GB", "B", "G", "diversity", "div.GB"), DEFAULT = "GB"}

\item{w.param}{in case div.GB selection mode is chosen}
}
\description{
Obtain id of the parents selected for the next generation based on the criteria chosen by the user
}
